First commited as before_9760_rev2

+++ src/main/java/net/spy/memcached/protocol/binary/MutatorOperationImpl.java
@@ -31,8 +31,8 @@ import net.spy.memcached.ops.OperationStatus;
 class MutatorOperationImpl extends SingleKeyOperationImpl implements
     MutatorOperation {
 
-  private static final int CMD_INCR = 0x05;
-  private static final int CMD_DECR = 0x06;
+  private static final byte CMD_INCR = 0x05;
+  private static final byte CMD_DECR = 0x06;
 
   private final Mutator mutator;
   private final long by;
@@ -70,7 +70,7 @@ class MutatorOperationImpl extends SingleKeyOperationImpl implements
         new OperationStatus(true, String.valueOf(decodeLong(pl, 0))));
   }
 
-  public int getBy() {
+  public long getBy() {
     return (int) by;
   }
